moved the ilogger to the libDomain and kept the consoleLogger implementation in the libUtilities
