+++ modules/libDomain/src/index.ts
@@ -8,3 +8,4 @@ export * from './entity_factory'
 export * from './ientity_state_repository'
 export * from './imessage_publisher'
 export * from './messages'
+export * from './ilogger'
