+++ modules/libInfrastructure/src/kafka_message_publisher.ts
@@ -5,7 +5,8 @@
 'use strict'
 
 // import * as kafka from 'kafka-node'
-import { ConsoleLogger, ILogger } from '@mojaloop-poc/lib-utilities'
+import { ConsoleLogger} from '@mojaloop-poc/lib-utilities'
+import { ILogger } from '@mojaloop-poc/lib-domain'
 import { IMessage, IMessagePublisher } from '@mojaloop-poc/lib-domain'
 import { KafkaGenericProducer } from './kafka_generic_producer'
