+++ modules/libUtilities/package.json
@@ -39,6 +39,7 @@
     "dep:update": "npx ncu -u"
   },
   "dependencies": {
+    "@mojaloop-poc/lib-domain": "^0.1.0"
   },
   "devDependencies": {
     "@types/kafka-node": "^3.0.0",
