+++ modules/participants/src/domain/participants_agg.ts
@@ -15,6 +15,7 @@ import { InvalidParticipantEvt } from '../messages/invalid_participant_evt'
 import { PayerFundsReservedEvt } from '../messages/payer_funds_reserved_evt'
 import { ParticipantCreatedEvt } from '../messages/participant_created_evt'
 import { NetCapLimitExceededEvt } from '../messages/netcaplimitexceeded_evt'
+import {ILogger} from '@mojaloop-poc/lib-domain'
 
 export enum ParticipantsAggTopics{
   'Commands' = 'ParticipantCommands',
@@ -23,8 +24,8 @@ export enum ParticipantsAggTopics{
 }
 
 export class ParticpantsAgg extends BaseAggregate<ParticipantEntity, ParticipantState> {
-  constructor (entityStateRepo: IEntityStateRepository<ParticipantState>, msgPublisher: IMessagePublisher) {
-    super(ParticipantsFactory.GetInstance(), entityStateRepo, msgPublisher)
+  constructor (entityStateRepo: IEntityStateRepository<ParticipantState>, msgPublisher: IMessagePublisher, logger:ILogger) {
+    super(ParticipantsFactory.GetInstance(), entityStateRepo, msgPublisher, logger)
     this._registerCommandHandler('CreateParticipantCmd', this.processCreateParticipantCommand)
     this._registerCommandHandler('ReservePayerFundsCmd', this.processReserveFundsCommand)
   }
