+++ modules/participants/src/infrastructure/redis_participant_repo.ts
@@ -6,7 +6,7 @@
 
 import * as redis from 'redis'
 import { IEntityStateRepository } from '@mojaloop-poc/lib-domain'
-import { ILogger } from '@mojaloop-poc/lib-utilities'
+import { ILogger } from '@mojaloop-poc/lib-domain'
 import { ParticipantState } from '../domain/participant_entity'
 
 export class RedisParticipantStateRepo implements IEntityStateRepository<ParticipantState> {
